item api jpa 방식으로 수정

+++ src/main/java/kr/co/gallery_jwt_jpa/item/ItemController.java
@@ -19,7 +19,7 @@ public class ItemController {
     private final ItemService itemService;
 
     @GetMapping
-    public ResponseEntity<?> readAll(@RequestParam(name = "id", required = false) List<Integer> ids) {
+    public ResponseEntity<?> readAll(@RequestParam(name = "id", required = false) List<Long> ids) {
         log.info("params: {}", ids);
         List<ItemGetRes> items = itemService.findAll(ids);
         return ResponseEntity.ok(items);
